Combine contract-related types in contract.go
